backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.0.2
pillow==9.5.0
opencv-python==4.7.0.72
numpy>=1.26.0
tensorflow==2.16.1
scikit-image==0.21.0
matplotlib==3.7.2
loguru==0.7.0
tensorflow-hub>=0.12.0
gdown>=4.6.0
dotenv==1.0.0